refactor(services): reorganize subgraph module into dedicated directory
Replace the scattered subgraph_impl and subgraph_service modules with a
consolidated subgraph directory structure for better organization and
maintainability.

Changes:
- Move subgraph_impl/http/ to subgraph/http/
- Move subgraph_service.rs to subgraph/service.rs
- Move subgraph.rs to subgraph/types.rs
- Create subgraph/mod.rs with proper re-exports
- Update all import statements throughout codebase
- Remove old subgraph_impl directory entirely

The new structure provides clearer separation of concerns with:
- types.rs: Type definitions and request/response structures
- service.rs: Core subgraph service implementation
- http/: HTTP client utilities and fetch logic

All existing public APIs are preserved through re-exports in mod.rs
ensuring backward compatibility.

Review focus: Verify that all subgraph-related functionality still
works correctly, especially WebSocket subscriptions, APQ handling,
and HTTP client behavior. Check that no imports were missed in the
reorganization.
