+++ apollo-router/src/services/subgraph/mod.rs
@@ -1,0 +1,15 @@
+//! Subgraph service modules
+
+pub(crate) mod http;
+pub(crate) mod service;
+pub(crate) mod types;
+
+// Re-export types for backward compatibility
+// Re-export service items for backward compatibility
+pub(crate) use service::APPLICATION_JSON_HEADER_VALUE;
+pub(crate) use service::MakeSubgraphService;
+pub(crate) use service::SubgraphService;
+pub(crate) use service::SubgraphServiceFactory;
+pub(crate) use service::generate_tls_client_config;
+pub(crate) use service::process_batches;
+pub use types::*;
