--- apollo-router/src/services/subgraph_impl/mod.rs
@@ -1,3 +1,0 @@
-//! Subgraph service modules
-
-pub(crate) mod http;
